Potential fix for code scanning alert no. 54: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger integer type to avoid overflow. This can be done by casting one of the operands to unsigned long before performing the multiplication. This way, the multiplication will be done using the larger type, and the result will be correct.

Specifically, we will cast nbstates to unsigned long before the multiplication on line 543. This change will ensure that the multiplication is performed using unsigned long, preventing overflow.

+++ libxml2-2.9.9/xmlregexp.c
@@ -540,7 +540,7 @@ xmlRegEpxFromParse(xmlRegParserCtxtPtr ctxt) {
 #ifdef DEBUG_COMPACTION
 	printf("Final: %d atoms\n", nbatoms);
 #endif
-	transitions = (int *) xmlMalloc((nbstates + 1) *
+	transitions = (int *) xmlMalloc(((unsigned long)(nbstates + 1)) *
 	                                (nbatoms + 1) * sizeof(int));
 	if (transitions == NULL) {
 	    xmlFree(stateRemap);
